BACKEND/requirements.txt
# Django Core
Django>=4.2.0

# Database
psycopg2-binary>=2.9.0

# REST API
djangorestframework>=3.14.0
drf-spectacular>=0.27.0

# CORS Support
django-cors-headers>=4.0.0

# Environment Variables
python-dotenv>=1.0.0

# Async Tasks
celery>=5.3.0
redis>=5.0.0

# Google Workspace Integration (Calendar, Meet, Drive)
google-api-python-client>=2.120.0
google-auth>=2.30.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0

